:bug: fix correctness issues in the kernels
The group-per-vertex mode seemed to be randomly crash, and our obersevations seem to indicate this is due to some race condition - particularly, some garbage community id is being selected as the best community, with weight 1, and this is causing the getModularity() function to crash. If the community keys are cleared, there is no crash, but the modularity of returned communities is bad.

The issue with group-per-vertex is now fixed. It was due to overlap of sketches for each group. That is, below should be p*SLOTS, not p*PLIM.

The issue with block-per-vertex approach with SHARED sketches seems to be that the threads are stepping over each other and causing atomicCAS() to fail. But, since there in no retry loop on the atomicCAS(), and the threads are mostly running synchronously, they have a tendency to cause significant failures. So, if we can keep track of separate sketches and merge them later, this would be the best option (maybe these slides will help). Else we need to have a retry loop with the shared sktech, which would badly affect performance.

So it looks like using a retry-loop solves the problem. Now we need to port it to the warp-optimized case too. Good! Next, we can go ahead and try a merging of independent sketches, but only SLOTS threads may be utilized during merging!